Add get_elasticsearch_index, move es utils to es.py.
